requirements.txt
APScheduler==3.6.3
asyncio==3.4.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.5
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-telegram-bot==13.15
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
six==1.16.0
tornado==6.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
